+++ CHANGELOG.rst
@@ -25,6 +25,7 @@ Changelog
 - Add ``nth_arg``. Thanks bharadwajyarlagadda_!
 - Add ``range_right``. Thanks bharadwajyarlagadda_!
 - Add ``subtract``. Thanks bharadwajyarlagadda_!
+- Add ``stub_false``. Thanks bharadwajyarlagadda_!
 - Add ``stub_string``. Thanks bharadwajyarlagadda_!
 - Add ``stub_true``. Thanks bharadwajyarlagadda_!
 - Add ``to_lower``. Thanks bharadwajyarlagadda_!